requirements.txt
requests
peakutils
pytest
pytest-mock
pandas
ebmdatalab-python
